fix: 폴더명 오타 수정 및usePreventScroll  훅 수정
