+++ components/Card/Card.tsx
@@ -7,7 +7,7 @@ import Image from "next/image";
 import AddModal from "@/components/Modals/AddModal/AddModal";
 import DeleteLinkModal from "@/components/Modals/DeleteLinkModal/DeleteLinkModal";
 import useOutsideClick from "@/hooks/useOutsideClick";
-import beautifyDate from "@/lib/beautifyData";
+import beautifyDate from "@/lib/beautifyDate";
 import { ILinkData } from "@/lib/getFolderData";
 
 import styles from "./Card.module.scss";
